refactor: move all keystone related files to keystone package
